cleanup, remove logout if refresh failed
