+++ extension/app/src/lib/storage.ts
@@ -1,5 +1,4 @@
 import type { ContentFragmentType, MeResponseType } from "@dust-tt/client";
-import { refreshToken } from "@extension/lib/auth";
 import type { Auth0AuthorizeResponse } from "@extension/lib/messages";
 import type { UploadedFileWithKind } from "@extension/lib/types";
 
@@ -46,18 +45,6 @@ export const getStoredTokens = async (): Promise<StoredTokens | null> => {
     };
   }
   return null;
-};
-
-export const getAccessToken = async (): Promise<string | null> => {
-  let tokens = await getStoredTokens();
-  if (!tokens || !tokens.accessToken || tokens.expiresAt < Date.now()) {
-    const refreshRes = await refreshToken();
-    if (refreshRes.isOk()) {
-      tokens = refreshRes.value;
-    }
-  }
-
-  return tokens?.accessToken ?? null;
 };
 
 type ConversationContext = {
