+++ extension/app/src/lib/swr.ts
@@ -1,5 +1,5 @@
 import { useAuthErrorCheck } from "@extension/hooks/useAuthErrorCheck";
-import { getAccessToken } from "@extension/lib/storage";
+import { getAccessToken } from "@extension/lib/auth";
 import { useCallback } from "react";
 import type { Fetcher, Key, SWRConfiguration } from "swr";
 import useSWR, { useSWRConfig } from "swr";
